client/cronjob/requirements.txt
python-dotenv==1.1.0
requests==2.31.0